app/services/ai_services/lora_trainer/requirements.txt
accelerate==0.25.0
transformers==4.36.2
diffusers[torch]==0.25.0
ftfy==6.1.1
opencv-python==4.7.0.68
einops==0.6.1
pytorch-lightning==1.9.0
tensorboard==2.10.1
safetensors==0.3.1
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.20.1
xformers
open-clip-torch==2.20.0